% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{common}
\alias{common}
\title{Common Docs}
\arguments{
\item{values}{A vector of values.}

\item{...}{Additional arguments, see \href{https://github.com/WebLogin/trendchart-elements#usage}{usage}}
}
\description{
Common Docs
}
