% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{style}
\alias{style}
\title{Style}
\usage{
style(...)
}
\arguments{
\item{...}{Named arguments}
}
\description{
Style a chart.
see \href{https://github.com/WebLogin/trendchart-elements#css-variables}{variables}
}
