% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-bar.R
\name{trend_bar}
\alias{trend_bar}
\title{Bar}
\usage{
trend_bar(values, ...)
}
\arguments{
\item{values}{A vector of values.}

\item{...}{Additional arguments, see \href{https://github.com/WebLogin/trendchart-elements#usage}{usage}}
}
\description{
Bar chart.
}
