% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-column.R
\name{trend_column}
\alias{trend_column}
\title{Column}
\usage{
trend_column(values, ...)
}
\arguments{
\item{values}{A vector of values.}

\item{...}{Additional arguments, see \href{https://github.com/WebLogin/trendchart-elements#usage}{usage}}
}
\description{
Column chart.
}
