% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-line.R
\name{trend_line}
\alias{trend_line}
\title{Line}
\usage{
trend_line(values, ...)
}
\arguments{
\item{values}{A vector of values.}

\item{...}{Additional arguments, see \href{https://github.com/WebLogin/trendchart-elements#usage}{usage}}
}
\description{
Line chart.
}
