% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts-pie.R
\name{trend_pie}
\alias{trend_pie}
\title{Pie}
\usage{
trend_pie(values, ...)
}
\arguments{
\item{values}{A vector of values.}

\item{...}{Additional arguments, see \href{https://github.com/WebLogin/trendchart-elements#usage}{usage}}
}
\description{
Pie chart.
}
